Add EventLike class and UniqueEvent
Ignore-this: 1bf7987aabb7958ca336e262d546de4b

This adds new event type called UniqueEvent. It is an Event which is supposed
to be unique for each step, that is, if a wire observes an UniqueEvent on its
input on two different ticks, it can assume that these are completely different
events. This enables one to, in addition to usual intervals-related event
handling, extract Maybe from an Event or apply a monadic action to it.

Some functions already produce UniqueEvents in fact (they were converted) --
such as 'now' or 'at'; other functions can produce any kind of events (like
'never'). All functions which worked with events were converted to work with any
kind of events. In addition, functions 'onUEventM' and 'fired' were added, which
work only with UniqueEvents.

The biggest problem of this patch is that we can't fully guarantee main
assumption of UniqueEvent with types; it can be broken with a wire that saves
incoming UniqueEvent and re-emits it later.

darcs-hash:20141205104309-eaee3-7ff10c07075e837d8b97e89a54f08542fa099f09